+++ examples/bdt_sklearn_to_tmva_Grad.py
@@ -31,7 +31,7 @@ X_train, y_train, w_train = X, y, w
 # Declare BDT - we are going to use AdaBoost Decision Tree
 bdt = GradientBoostingClassifier(
     n_estimators=200, learning_rate=0.5,
-    min_samples_leaf=0.05*len(X_train),
+    min_samples_leaf=int(0.05*len(X_train)),
     max_depth=3, random_state=0)
 
 # Train BDT
